SECOPS/whats_new/requirements.txt
requests
Flask
flask-cors
apscheduler
google-generativeai
